requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
aioredis==1.3.1
alabaster==0.7.12
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
bleach==3.1.5
blessings==1.7
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorful==0.5.4
coverage==5.2.1
cryptography==3.0
cycler==0.10.0
docutils==0.16
filelock==3.0.12
gast==0.3.3
google==3.0.0
google-api-core==1.22.1
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
graphviz==0.14.1
grpcio==1.31.0
h5py==2.10.0
hiredis==1.1.0
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
iniconfig==1.0.1
jeepney==0.4.3
Jinja2==2.11.2
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
keyring==21.3.0
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
more-itertools==8.4.0
msgpack==1.0.0
multidict==4.7.6
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.10
opencensus-context==0.1.1
opt-einsum==3.3.0
packaging==20.4
Pillow==7.2.0
pipdeptree==1.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.8.0
protobuf==3.13.0
psutil==5.7.2
py==1.9.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.16.0
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
ray==0.8.7
readme-renderer==26.0
redis==3.4.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
scipy==1.4.1
SecretStorage==3.1.2
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
tqdm==4.48.2
twine==3.2.0
typing-extensions==3.7.4.2
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.1.0
